refactor: split project picker components
