+++ Muxy/Views/Components/ProjectPickerOverlay.swift
@@ -1,5 +1,4 @@
 import AppKit
-import Carbon.HIToolbox
 import SwiftUI
 
 struct ProjectPickerOverlay: View {
@@ -420,37 +419,6 @@ private enum ProjectPickerDirectoryLoadState: Equatable {
     }
 }
 
-struct ProjectPickerDirectorySnapshot {
-    let rows: [String]
-    let readFailed: Bool
-
-    static func load(navigator: ProjectPickerNavigator) -> ProjectPickerDirectorySnapshot {
-        do {
-            let urls = try FileManager.default.contentsOfDirectory(
-                at: URL(fileURLWithPath: navigator.directoryPath),
-                includingPropertiesForKeys: [.isDirectoryKey, .isSymbolicLinkKey],
-                options: []
-            )
-            let names = urls.compactMap { url -> String? in
-                guard isDirectoryOrDirectorySymlink(url) else { return nil }
-                return url.lastPathComponent
-            }
-            return ProjectPickerDirectorySnapshot(rows: navigator.directoryRows(from: names), readFailed: false)
-        } catch {
-            let rows = navigator.directoryPath == "/" ? [] : [ProjectPickerNavigator.parentDirectoryRow]
-            return ProjectPickerDirectorySnapshot(rows: rows, readFailed: true)
-        }
-    }
-
-    private static func isDirectoryOrDirectorySymlink(_ url: URL) -> Bool {
-        let values = try? url.resourceValues(forKeys: [.isDirectoryKey, .isSymbolicLinkKey])
-        if values?.isDirectory == true { return true }
-        guard values?.isSymbolicLink == true else { return false }
-        var isDirectory = ObjCBool(false)
-        return FileManager.default.fileExists(atPath: url.path, isDirectory: &isDirectory) && isDirectory.boolValue
-    }
-}
-
 private struct ProjectPickerDirectoryRow: View {
     let row: String
     let isParent: Bool
@@ -477,179 +445,5 @@ private struct ProjectPickerDirectoryRow: View {
         .background(isHighlighted ? MuxyTheme.surface : hovered ? MuxyTheme.hover : .clear)
         .contentShape(Rectangle())
         .onHover { hovered = $0 }
-    }
-}
-
-private struct ProjectPickerShortcutHint: View {
-    let keycap: ProjectPickerShortcutKeycap
-    let label: String
-
-    var body: some View {
-        HStack(spacing: UIMetrics.scaled(4)) {
-            HStack(spacing: UIMetrics.scaled(3)) {
-                ForEach(Array(keycap.parts.enumerated()), id: \.offset) { _, part in
-                    keycapPart(part)
-                }
-            }
-            .lineLimit(1)
-            .fixedSize(horizontal: true, vertical: false)
-            .padding(.horizontal, UIMetrics.scaled(4))
-            .padding(.vertical, UIMetrics.scaled(2))
-            .background(MuxyTheme.surface, in: RoundedRectangle(cornerRadius: UIMetrics.radiusSM))
-            .overlay(RoundedRectangle(cornerRadius: UIMetrics.radiusSM).stroke(MuxyTheme.border, lineWidth: 1))
-            Text(label)
-                .font(.system(size: UIMetrics.fontFootnote, weight: .medium))
-                .foregroundStyle(MuxyTheme.fgDim)
-                .lineLimit(1)
-                .fixedSize(horizontal: true, vertical: false)
-        }
-        .fixedSize(horizontal: true, vertical: false)
-    }
-
-    @ViewBuilder
-    private func keycapPart(_ part: ProjectPickerShortcutKeycapPart) -> some View {
-        switch part {
-        case let .symbol(name):
-            Image(systemName: name)
-                .font(.system(size: UIMetrics.fontCaption, weight: .semibold))
-                .foregroundStyle(MuxyTheme.fgMuted)
-        case let .text(text):
-            Text(text)
-                .font(.system(size: UIMetrics.fontCaption, weight: .semibold, design: .monospaced))
-                .foregroundStyle(MuxyTheme.fgMuted)
-        }
-    }
-}
-
-private struct ProjectPickerShortcutKeycap: Hashable {
-    let parts: [ProjectPickerShortcutKeycapPart]
-
-    static let navigate = ProjectPickerShortcutKeycap(parts: [.symbol("arrow.up"), .symbol("arrow.down")])
-    static let returnKey = ProjectPickerShortcutKeycap(parts: [.text("Return")])
-    static let commandReturn = ProjectPickerShortcutKeycap(parts: [.symbol("command"), .text("Return")])
-    static let escape = ProjectPickerShortcutKeycap(parts: [.text("Esc")])
-    static let optionDelete = ProjectPickerShortcutKeycap(parts: [.symbol("option"), .symbol("delete.left")])
-}
-
-private enum ProjectPickerShortcutKeycapPart: Hashable {
-    case symbol(String)
-    case text(String)
-}
-
-private struct ProjectPickerPathField: NSViewRepresentable {
-    @Binding var text: String
-    let onSubmit: () -> Void
-    let onCommandSubmit: () -> Void
-    let onEscape: () -> Void
-    let onArrowUp: () -> Void
-    let onArrowDown: () -> Void
-    let onTab: () -> Void
-    let onGoUp: () -> Void
-
-    func makeCoordinator() -> Coordinator {
-        Coordinator(parent: self)
-    }
-
-    func makeNSView(context: Context) -> NSTextField {
-        let field = ProjectPickerNSTextField()
-        field.delegate = context.coordinator
-        field.isBordered = false
-        field.drawsBackground = false
-        field.focusRingType = .none
-        field.font = .monospacedSystemFont(ofSize: UIMetrics.fontEmphasis, weight: .regular)
-        field.textColor = NSColor(MuxyTheme.fg)
-        field.stringValue = text
-        field.onEscape = onEscape
-        field.onCommandSubmit = onCommandSubmit
-        DispatchQueue.main.async {
-            field.window?.makeFirstResponder(field)
-            field.moveCursorToEnd()
-        }
-        return field
-    }
-
-    func updateNSView(_ nsView: NSTextField, context: Context) {
-        context.coordinator.parent = self
-        if nsView.stringValue != text {
-            nsView.stringValue = text
-        }
-        if let field = nsView as? ProjectPickerNSTextField {
-            field.onEscape = onEscape
-            field.onCommandSubmit = onCommandSubmit
-        }
-    }
-
-    @MainActor
-    final class Coordinator: NSObject, NSTextFieldDelegate {
-        var parent: ProjectPickerPathField
-
-        init(parent: ProjectPickerPathField) {
-            self.parent = parent
-        }
-
-        func controlTextDidChange(_ obj: Notification) {
-            guard let field = obj.object as? NSTextField else { return }
-            parent.text = field.stringValue
-        }
-
-        func control(_ control: NSControl, textView: NSTextView, doCommandBy commandSelector: Selector) -> Bool {
-            if commandSelector == #selector(NSResponder.insertNewline(_:)) {
-                parent.onSubmit()
-                return true
-            }
-            if commandSelector == #selector(NSResponder.insertTab(_:)) {
-                parent.onTab()
-                return true
-            }
-            if commandSelector == #selector(NSResponder.moveUp(_:)) {
-                parent.onArrowUp()
-                return true
-            }
-            if commandSelector == #selector(NSResponder.moveDown(_:)) {
-                parent.onArrowDown()
-                return true
-            }
-            if commandSelector == #selector(NSResponder.deleteWordBackward(_:)), shouldGoUpOnDeleteBackward(textView) {
-                parent.onGoUp()
-                return true
-            }
-            if commandSelector == #selector(NSResponder.deleteBackward(_:)), shouldGoUpOnDeleteBackward(textView) {
-                parent.onGoUp()
-                return true
-            }
-            return false
-        }
-
-        private func shouldGoUpOnDeleteBackward(_ textView: NSTextView) -> Bool {
-            let selectedRange = textView.selectedRange()
-            guard selectedRange.length == 0, selectedRange.location == textView.string.utf16.count else { return false }
-            let value = textView.string.trimmingCharacters(in: .whitespacesAndNewlines)
-            return value.isEmpty || value == "~" || value.hasSuffix("/")
-        }
-    }
-}
-
-private final class ProjectPickerNSTextField: NSTextField {
-    var onEscape: (() -> Void)?
-    var onCommandSubmit: (() -> Void)?
-
-    func moveCursorToEnd() {
-        guard let editor = currentEditor() else { return }
-        editor.selectedRange = NSRange(location: stringValue.utf16.count, length: 0)
-    }
-
-    override func performKeyEquivalent(with event: NSEvent) -> Bool {
-        guard window?.firstResponder === currentEditor() else {
-            return super.performKeyEquivalent(with: event)
-        }
-        if event.keyCode == kVK_Escape {
-            onEscape?()
-            return true
-        }
-        if event.keyCode == kVK_Return, event.modifierFlags.contains(.command) {
-            onCommandSubmit?()
-            return true
-        }
-        return super.performKeyEquivalent(with: event)
     }
 }
